refactor: 접근제어자 변경 및 Exception 클래스 exception 패키지로 이동
