+++ src/main/java/racingcar/domain/CarValidator.java
@@ -1,5 +1,7 @@
 package racingcar.domain;
 
+import racingcar.domain.exception.InvalidCarNameLengthException;
+
 class CarValidator {
 
     static void validCarLength(String name) {
