+++ src/main/java/racingcar/domain/RoundResults.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class RoundResults {
     private final List<RoundResult> roundResults;
 
-    public RoundResults(List<RoundResult> roundResults) {
+    RoundResults(List<RoundResult> roundResults) {
         this.roundResults = roundResults;
     }
