+++ src/main/java/racingcar/domain/exception/NotExistMovedCarException.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.exception;
 
 public class NotExistMovedCarException extends RuntimeException {
     private static final String ERROR_MESSAGE = "경주한 차가 존재하지 않습니다.";
