+++ src/main/java/racingcar/view/InputValidator.java
@@ -1,5 +1,7 @@
 package racingcar.view;
 
+import racingcar.view.exception.NegativeRoundsException;
+
 import java.util.regex.Pattern;
 
 class InputValidator {
